+++ src/main/kotlin/blackjack/model/Card.kt
@@ -1,3 +1,3 @@
-package blackjack
+package blackjack.model
 
 data class Card(private val shape: CardShape, val number: CardNumber)
