+++ src/main/kotlin/blackjack/model/CardHandState.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 enum class CardHandState {
     BLACKJACK,
