+++ src/main/kotlin/blackjack/model/CardNumber.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 enum class CardNumber(
     val number: Int,
