+++ src/main/kotlin/blackjack/model/CardShape.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.model
 
 enum class CardShape(
     name: String,
