--- src/test/kotlin/blackjack/BlackJackTest.kt
@@ -1,3 +1,0 @@
-package blackjack
-
-class BlackJackTest
