+++ src/test/kotlin/blackjack/model/BlackJackTest.kt
@@ -1,0 +1,14 @@
+package blackjack.model
+
+import org.junit.jupiter.api.Test
+
+class BlackJackTest {
+    @Test
+    fun `카드 핸드 상태를 가지고 새 카드를 뽑는다`() {
+        // 플레이어
+
+        // val state = 플레이어.getState()
+
+        // player.method(state)
+    }
+}
